Add files via upload
This is basically a coursera project based course. And I completed it and earned certification. This project was mainly on the feed forward neural network and hands-on-experience on Keras (tensorflow at backend) to classify the Reuters topic dataset. This dataset contains 46 unique class labels.  As for input I am using a matrix form of dataset and representing a single training or test example with One-Hot encoding. And for multiple classification problems I used 'softmax' which results in probabilities of these 46 classes.